refactor: entity와 dao 패키지 상세화
