+++ src/main/java/com/wanted/domain/restaurant/dao/employee/RestaurantEmployeeRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.restaurant.dao;
+package com.wanted.domain.restaurant.dao.employee;
 
-import com.wanted.domain.restaurant.entity.RestaurantEmployee;
+import com.wanted.domain.restaurant.entity.employee.RestaurantEmployee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantEmployeeRepository extends JpaRepository<RestaurantEmployee, Long> {
