+++ src/main/java/com/wanted/domain/restaurant/dao/facility/RestaurantFacilityRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.restaurant.dao;
+package com.wanted.domain.restaurant.dao.facility;
 
-import com.wanted.domain.restaurant.entity.RestaurantFacility;
+import com.wanted.domain.restaurant.entity.facility.RestaurantFacility;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantFacilityRepository extends JpaRepository<RestaurantFacility, Long> {
