+++ src/main/java/com/wanted/domain/restaurant/dao/hygiene/RestaurantHygieneRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.restaurant.dao;
+package com.wanted.domain.restaurant.dao.hygiene;
 
-import com.wanted.domain.restaurant.entity.RestaurantHygiene;
+import com.wanted.domain.restaurant.entity.hygiene.RestaurantHygiene;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantHygieneRepository extends JpaRepository<RestaurantHygiene, Long> {
