+++ src/main/java/com/wanted/domain/restaurant/dao/location/RestaurantLocationRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.restaurant.dao;
+package com.wanted.domain.restaurant.dao.location;
 
-import com.wanted.domain.restaurant.entity.RestaurantLocation;
+import com.wanted.domain.restaurant.entity.location.RestaurantLocation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantLocationRepository extends JpaRepository<RestaurantLocation, Long> {
