+++ src/main/java/com/wanted/domain/restaurant/dao/site/RestaurantSiteRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.restaurant.dao;
+package com.wanted.domain.restaurant.dao.site;
 
-import com.wanted.domain.restaurant.entity.RestaurantSite;
+import com.wanted.domain.restaurant.entity.site.RestaurantSite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantSiteRepository extends JpaRepository<RestaurantSite, Long> {
