+++ src/main/java/com/wanted/domain/restaurant/dao/type/RestaurantTypeRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.restaurant.dao;
+package com.wanted.domain.restaurant.dao.type;
 
-import com.wanted.domain.restaurant.entity.RestaurantType;
+import com.wanted.domain.restaurant.entity.type.RestaurantType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantTypeRepository extends JpaRepository<RestaurantType, Long> {
