+++ src/main/java/com/wanted/domain/restaurant/dao/workplace/RestaurantWorkplaceRepository.java
@@ -1,6 +1,6 @@
-package com.wanted.domain.restaurant.dao;
+package com.wanted.domain.restaurant.dao.workplace;
 
-import com.wanted.domain.restaurant.entity.RestaurantWorkplace;
+import com.wanted.domain.restaurant.entity.workplace.RestaurantWorkplace;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RestaurantWorkplaceRepository extends JpaRepository<RestaurantWorkplace, Long> {
