+++ src/main/java/com/wanted/domain/restaurant/entity/hygiene/RestaurantHygiene.java
@@ -1,4 +1,6 @@
-package com.wanted.domain.restaurant.entity;
+package com.wanted.domain.restaurant.entity.hygiene;
+
+import com.wanted.domain.restaurant.entity.Restaurant;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
@@ -9,40 +11,31 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
-import java.time.LocalDate;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 맛집의 사업장 정보를 담는 엔티티
+ * 맛집의 위생 정보를 담는 엔티티
  */
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class RestaurantWorkplace {
+public class RestaurantHygiene {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "restaurant_workplace_id")
+  @Column(name = "restaurant_hygiene_id")
   private Long id;
 
-  // 사업장명
+  // 위생 업종명
   @Column(nullable = false)
-  private String workplaceName;
+  private String industryName;
 
-  // 사업자 인허가 일자
+  // 위생 업태명
   @Column(nullable = false)
-  private LocalDate licenseDate;
-
-  // 사업자 상태명
-  @Column(nullable = false)
-  private String businessStatus;
-
-  // 영업장 주변 구분명
-  @Column(nullable = false)
-  private String classificationName;
+  private String businessName;
 
   // Restaurnt 테이블이 부모 테이블
   @ManyToOne(fetch = FetchType.LAZY)
@@ -50,12 +43,9 @@ public class RestaurantWorkplace {
   private Restaurant restaurant;
 
   @Builder
-  private RestaurantWorkplace(String workplaceName, LocalDate licenseDate, String businessStatus,
-      String classificationName, Restaurant restaurant) {
-    this.workplaceName = workplaceName;
-    this.licenseDate = licenseDate;
-    this.businessStatus = businessStatus;
-    this.classificationName = classificationName;
+  private RestaurantHygiene(String industryName, String businessName, Restaurant restaurant) {
+    this.industryName = industryName;
+    this.businessName = businessName;
     this.restaurant = restaurant;
   }
 }
