+++ src/main/java/com/wanted/domain/restaurant/entity/site/RestaurantSite.java
@@ -1,11 +1,10 @@
-package com.wanted.domain.restaurant.entity;
+package com.wanted.domain.restaurant.entity.site;
 
-import com.wanted.domain.restaurant.constant.FoodType;
+import com.wanted.domain.restaurant.entity.Restaurant;
+
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -18,22 +17,33 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 맛집의 타입 정보를 담는 엔티티
+ * 맛집의 소재지 정보를 담는 엔티티
  */
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class RestaurantType {
+public class RestaurantSite {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "restaurant_type_id")
+  @Column(name = "restaurant_site_id")
   private Long id;
 
-  // 맛집 타입 (중식, 일식, 패스트푸드)
+  // 소재지 도로명 주소
   @Column(nullable = false)
-  @Enumerated(EnumType.STRING)
-  private FoodType type;
+  private String roadNameAddress;
+
+  // 소재지 지번 주소
+  @Column(nullable = false)
+  private String lotNumberAddress;
+
+  // 소재지 우편번호
+  @Column(nullable = false)
+  private Integer zipCode;
+
+  // 소재지 면적
+  @Column(nullable = false)
+  private Long area;
 
   // Restaurnt 테이블이 부모 테이블
   @ManyToOne(fetch = FetchType.LAZY)
@@ -41,8 +51,12 @@ public class RestaurantType {
   private Restaurant restaurant;
 
   @Builder
-  private RestaurantType(FoodType type, Restaurant restaurant) {
-    this.type = type;
+  private RestaurantSite(String roadNameAddress, String lotNumberAddress, Integer zipCode, Long area,
+      Restaurant restaurant) {
+    this.roadNameAddress = roadNameAddress;
+    this.lotNumberAddress = lotNumberAddress;
+    this.zipCode = zipCode;
+    this.area = area;
     this.restaurant = restaurant;
   }
 }
