+++ src/main/java/com/wanted/domain/restaurant/entity/type/RestaurantType.java
@@ -1,8 +1,13 @@
-package com.wanted.domain.restaurant.entity;
+package com.wanted.domain.restaurant.entity.type;
+
+import com.wanted.domain.restaurant.constant.FoodType;
+import com.wanted.domain.restaurant.entity.Restaurant;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -15,25 +20,22 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 맛집의 위생 정보를 담는 엔티티
+ * 맛집의 타입 정보를 담는 엔티티
  */
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-public class RestaurantHygiene {
+public class RestaurantType {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "restaurant_hygiene_id")
+  @Column(name = "restaurant_type_id")
   private Long id;
 
-  // 위생 업종명
+  // 맛집 타입 (중식, 일식, 패스트푸드)
   @Column(nullable = false)
-  private String industryName;
-
-  // 위생 업태명
-  @Column(nullable = false)
-  private String businessName;
+  @Enumerated(EnumType.STRING)
+  private FoodType type;
 
   // Restaurnt 테이블이 부모 테이블
   @ManyToOne(fetch = FetchType.LAZY)
@@ -41,9 +43,8 @@ public class RestaurantHygiene {
   private Restaurant restaurant;
 
   @Builder
-  private RestaurantHygiene(String industryName, String businessName, Restaurant restaurant) {
-    this.industryName = industryName;
-    this.businessName = businessName;
+  private RestaurantType(FoodType type, Restaurant restaurant) {
+    this.type = type;
     this.restaurant = restaurant;
   }
 }
