+++ src/main/java/com/wanted/domain/restaurant/openapi/dao/RestaurantDataDBInserter.java
@@ -1,20 +1,20 @@
 package com.wanted.domain.restaurant.openapi.dao;
 
 import com.wanted.domain.restaurant.constant.FoodType;
-import com.wanted.domain.restaurant.dao.RestaurantEmployeeRepository;
-import com.wanted.domain.restaurant.dao.RestaurantFacilityRepository;
-import com.wanted.domain.restaurant.dao.RestaurantHygieneRepository;
+import com.wanted.domain.restaurant.dao.employee.RestaurantEmployeeRepository;
+import com.wanted.domain.restaurant.dao.facility.RestaurantFacilityRepository;
+import com.wanted.domain.restaurant.dao.hygiene.RestaurantHygieneRepository;
 import com.wanted.domain.restaurant.dao.RestaurantRepository;
-import com.wanted.domain.restaurant.dao.RestaurantSiteRepository;
-import com.wanted.domain.restaurant.dao.RestaurantTypeRepository;
-import com.wanted.domain.restaurant.dao.RestaurantWorkplaceRepository;
+import com.wanted.domain.restaurant.dao.site.RestaurantSiteRepository;
+import com.wanted.domain.restaurant.dao.type.RestaurantTypeRepository;
+import com.wanted.domain.restaurant.dao.workplace.RestaurantWorkplaceRepository;
 import com.wanted.domain.restaurant.entity.Restaurant;
-import com.wanted.domain.restaurant.entity.RestaurantEmployee;
-import com.wanted.domain.restaurant.entity.RestaurantFacility;
-import com.wanted.domain.restaurant.entity.RestaurantHygiene;
-import com.wanted.domain.restaurant.entity.RestaurantSite;
-import com.wanted.domain.restaurant.entity.RestaurantType;
-import com.wanted.domain.restaurant.entity.RestaurantWorkplace;
+import com.wanted.domain.restaurant.entity.employee.RestaurantEmployee;
+import com.wanted.domain.restaurant.entity.facility.RestaurantFacility;
+import com.wanted.domain.restaurant.entity.hygiene.RestaurantHygiene;
+import com.wanted.domain.restaurant.entity.site.RestaurantSite;
+import com.wanted.domain.restaurant.entity.type.RestaurantType;
+import com.wanted.domain.restaurant.entity.workplace.RestaurantWorkplace;
 import com.wanted.domain.restaurant.openapi.dto.RestaurantOpenApiData;
 import java.time.LocalDate;
 import java.util.ArrayList;
